backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
psycopg[binary]==3.2.3
boto3==1.35.92
bcrypt==4.2.1
PyJWT==2.10.1
email-validator==2.2.0
pillow==10.4.0
slowapi==0.1.9
